pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm_defender"
version = "0.9.3"
description = "This project implements the llm-defender subnet for Bittensor"
authors = [
    { name = "synapsec.ai", email = "contact@synapsec.ai" }
]
readme = { file = "README.md", content-type = "text/markdown" }
license = { file = "LICENSE" }
classifiers = [
    "Development Status :: 3 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
requires-python = ">=3.10,<3.12"

dependencies = [
    "bittensor==7.3.1",
    "python-dotenv==1.0.1",
]

[project.urls]
homepage = "https://github.com/synapsec-ai/llm-defender-subnet"

[project.optional-dependencies]
wandb = ["wandb==0.17.0"]
miner = ["transformers==4.41.2", "torch==2.3.0"]
validator = ["scikit-learn==1.5.0", "numpy==1.26.4", "english-words==2.0.1", "openai==1.34.0", "datasets==2.20.0", "pysha3==1.0.2", "wonderwords==2.2.0"]
api = ["fastapi==0.110.1", "uvicorn==0.22.0", "python-multipart==0.0.9"]

[tool.setuptools.packages.find]
include = ["llm_defender"]
